Breaking ExecutorWithExceptionHandling decorator into separate implementation of Executor
