+++ osbenchmark/builder/executors/executor.py
@@ -1,16 +1,13 @@
 import logging
 
-from osbenchmark.exceptions import ExecutorError
-
 
 class Executor:
     """
     Executors are used to run shell commands on the cluster hosts. Implementations of this class will use various
     technologies to interface with the hosts of a cluster.
     """
-    def __init__(self, executor):
+    def __init__(self):
         self.logger = logging.getLogger(__name__)
-        self.executor = executor
 
     def execute(self, host, command, **kwargs):
         """
@@ -20,12 +17,9 @@ class Executor:
         ;param command: A shell command as a string
         ;return output: The output of the command
         """
-        try:
-            self.executor.execute(host, command, kwargs)
-        except Exception as e:
-            raise ExecutorError("Command \"{}\" on host \"{}\" failed to execute".format(command, host), e)
+        raise NotImplementedError
 
-    def copy(self, host, source, destination, **kwargs):
+    def copy(self, host, source, destination):
         """
         Copies the specified source to the destination on the given host
 
@@ -34,7 +28,4 @@ class Executor:
         ;param destination: The destination to copy to
         ;return None
         """
-        try:
-            self.executor.copy(host, source, destination, kwargs)
-        except Exception as e:
-            raise ExecutorError("Copying \"{}\" to \"{}\" failed on host \"{}\"".format(source, destination, host), e)
+        raise NotImplementedError
